requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

# Data Processing & Augmentation
albumentations>=1.3.1
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0

# Metrics & Visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.65.0

# Web Application
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
joblib>=1.3.0
